@@ -19,6 +19,7 @@ import { useMounted } from 'core/useMounted';
 import { ActionType } from 'data/actionTypes';
 import { useCancellation } from 'core/useCancellation';
 import { getItemAsync, setItem, saveAsync } from 'data/config/widget/actions';
+import { useCleaning } from 'core/useCleaning';
 
 const schema = yup.object().shape({
   name: yup
@@ -32,47 +33,60 @@ const schema = yup.object().shape({
 
 export const WidgetItem: React.FC = () => {
   useCancellation(ActionType.CONFIG_WIDGET_SAVEASYNC);
+  useCleaning(setItem);
   const match = useMatch<{ id: number }>();
   const dispatch = useDispatch();
-  const mounted = useMounted(useCallback(() => dispatch(setItem()), [dispatch]));
   const item = useSelector((state: StoreType) => state.config.widget.item);
   const [messages, setMessages] = useState<string | string[]>();
   const get = useCallback(() => dispatch(getItemAsync(match.params)), [dispatch, match.params]);
   const save = useCallback(async () => {
-    setMessages(undefined);
-    dispatch(saveAsync(item));
-    if (mounted.current) {
-      // if (result.isError) setMessages(result.error);
-      // else history.goBack();
-    }
-  }, [dispatch, item, mounted]);
+    schema
+      .validate(item)
+      .then(() => {
+        dispatch(saveAsync(item));
+        setMessages(undefined);
+      })
+      .catch(x => setMessages(x.message));
+  }, [dispatch, item]);
   useEffect(() => {
     if (match.params.id > 0) get();
   }, [get, match.params]);
   const change = (field: string, value: any) => {
     dispatch(setItem({ ...item, [field]: value }));
   };
-  console.log('***', mounted, match);
   return (
     <DefaultPage title="Widget Item Test">
       <RepeatPanel actionType={ActionType.CONFIG_WIDGET_GETITEMASYNC} action={get}>
-        <MessagesView messages={messages}></MessagesView>
+        <MessagesView messages={messages} actionType={ActionType.CONFIG_WIDGET_SAVEASYNC}></MessagesView>
         <Row>
-          <TextBoxField data={item} field="name" size={6} onChange={change} v={schema}>
+          <TextBoxField
+            name="name"
+            value={item.name}
+            size={6}
+            onChange={v => change('name', v)}
+            schema={schema}
+            fieldPath="name">
             Name
           </TextBoxField>
-          <SelectField data={item} field="type" values={WidgetType.all} size={6} onChange={change} v={schema}>
+          <SelectField
+            name="type"
+            value={item.type.toString()}
+            options={WidgetType.map}
+            getLabel={x => x}
+            size={6}
+            onChange={v => change('type', parseInt(v))}>
             Type
           </SelectField>
-          <TextBoxField data={item} field="parameters" size={6} onChange={change} v={schema}>
+          <TextBoxField name="parameters" value={item.parameters} size={6} onChange={v => change('parameters', v)}>
             Parameters
           </TextBoxField>
         </Row>
-        <div></div>
-        <LoadingButton actionType={ActionType.CONFIG_WIDGET_SAVEASYNC} primary onClick={save}>
-          Save
-        </LoadingButton>
-        <CancelButton ml="3" />
+        <div>
+          <LoadingButton actionType={ActionType.CONFIG_WIDGET_SAVEASYNC} primary onClick={save}>
+            Save
+          </LoadingButton>
+          <CancelButton ml="3" />
+        </div>
       </RepeatPanel>
     </DefaultPage>
   );
