@@ -7,7 +7,7 @@ import { LoginModel, AuthInfo } from './models';
 const baseUrl = 'auth';
 
 export function login(model: LoginModel) {
-  return http
-    .post<AuthInfo>(baseUrl, model)
-    .pipe(map<AuthInfo, AuthInfo>(x => ({ ...x, expires: x.expires != null ? new Date(x.expires) : null })));
+  return http.post<AuthInfo>(baseUrl, model).pipe(
+    map<AuthInfo, AuthInfo>(x => ({ ...x, expires: x.expires != null ? new Date(x.expires) : null }))
+  );
 }
