@@ -10,15 +10,17 @@ import { WidgetModel, WidgetModelEdit } from './models';
 const baseUrl = 'config/widget';
 
 export function getPage(page: number, widgetType?: number) {
-  return http.get<Page<WidgetModel>>(baseUrl, { page, type: widgetType }).pipe(
-    map(x => {
-      x.items.forEach(x => {
-        x.created = DateTime.parse(x.created as any);
-        x.updated = DateTime.parse(x.updated as any);
-      });
-      return x;
-    })
-  );
+  return http
+    .get<Page<WidgetModel>>(baseUrl, { page, type: widgetType })
+    .pipe(
+      map(x => {
+        x.items.forEach(x => {
+          x.created = DateTime.parse(x.created as any);
+          x.updated = DateTime.parse(x.updated as any);
+        });
+        return x;
+      })
+    );
 }
 
 export function getItem(id: number) {
